requirements.txt
Import Requirements

requests
pandas
typing
time
datetime
matplotlib.pyplot
datetime